pyproject.toml
[project]
name = "spotifications"
version = "0.1.0"
description = "Spotify crawler to get latests releases of your favorite artists"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "python-dotenv>=1.1.0",
    "spotipy>=2.25.1",
]
